bye config
