+++ src/__init__.py
@@ -1,0 +1,4 @@
+from .toolbox import load_var_file, ask_yes_no, set_var
+from .shared import compare_two_files
+
+__all__ = ['load_var_file', 'ask_yes_no', 'set_var', 'compare_two_files']
